rest of codes

+++ TozawaMaui/src/ShareRazorClassLibrary/Helpers/AuthStateProvider.cs
@@ -9,13 +9,18 @@ using Microsoft.JSInterop;
 using System.Globalization;
 using Blazored.LocalStorage;
 using ShareRazorClassLibrary.Configurations;
+using ShareRazorClassLibrary.Services;
+using System.IdentityModel.Tokens.Jwt;
+using Microsoft.IdentityModel.Tokens;
+using System.Text;
 
 namespace ShareRazorClassLibrary.Helpers;
 
-public class AuthStateProvider(ISessionStorageService sessionStorageService, ILocalStorageService localStorage, AppSettings appSettings) : AuthenticationStateProvider
+public class AuthStateProvider(ISessionStorageService sessionStorageService, ILocalStorageService localStorage, AppSettings appSettings, FirsloadState firsloadState) : AuthenticationStateProvider
 {
     private readonly ISessionStorageService _sessionStorageService = sessionStorageService;
     private readonly ILocalStorageService _localStorage = localStorage;
+    private readonly FirsloadState _firsloadState = firsloadState;
     private readonly AuthenticationState _anonymous = new(new ClaimsPrincipal(new ClaimsIdentity()));
     private readonly AppSettings _appSettings = appSettings;
 
@@ -23,6 +28,11 @@ public class AuthStateProvider(ISessionStorageService sessionStorageService, ILo
 
     public override async Task<AuthenticationState> GetAuthenticationStateAsync()
     {
+        if (!_firsloadState.IsFirstLoaded)
+        {
+            return new AuthenticationState(new ClaimsPrincipal());
+        }
+
         try
         {
             string token = null;
@@ -66,6 +76,10 @@ public class AuthStateProvider(ISessionStorageService sessionStorageService, ILo
     }
     public async Task RemoveCurrentUser()
     {
+        if (!_firsloadState.IsFirstLoaded)
+        {
+            return;
+        }
         if (await _sessionStorageService.ContainKeyAsync("currentUser"))
         {
             await _sessionStorageService.RemoveItemAsync("currentUser");
@@ -73,6 +87,10 @@ public class AuthStateProvider(ISessionStorageService sessionStorageService, ILo
     }
     public async Task SetCurrentUser(CurrentUserDto user)
     {
+        if (!_firsloadState.IsFirstLoaded)
+        {
+            return;
+        }
         if (await _sessionStorageService.ContainKeyAsync("currentUser"))
         {
             await _sessionStorageService.RemoveItemAsync("currentUser");
@@ -81,6 +99,10 @@ public class AuthStateProvider(ISessionStorageService sessionStorageService, ILo
     }
     public async void NotifyUserAuthentication()
     {
+        if (!_firsloadState.IsFirstLoaded)
+        {
+            return;
+        }
         var token = await _localStorage.GetItemAsync<string>("authToken");
 
         var refreshAt = DateTimeOffset.UtcNow.AddSeconds(Convert.ToDouble(_appSettings.JWTSettings.ExpiryInMinutes)).ToString(CultureInfo.InvariantCulture);
@@ -104,8 +126,37 @@ public class AuthStateProvider(ISessionStorageService sessionStorageService, ILo
         NotifyAuthenticationStateChanged(authState);
         UserAuthenticationChanged(this, new EventArgs());
     }
+    public bool ValidateCurrentToken(string token)
+    {
+        var myIssuer = _appSettings.JWTSettings.ValidIssuer;
+        var myAudience = _appSettings.JWTSettings.ValidAudience;
+
+        var tokenHandler = new JwtSecurityTokenHandler();
+        try
+        {
+            tokenHandler.ValidateToken(token, new TokenValidationParameters
+            {
+                ValidateIssuerSigningKey = true,
+                ValidateIssuer = true,
+                ValidateAudience = true,
+                ValidIssuer = myIssuer,
+                ValidAudience = myAudience,
+                IssuerSigningKey = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(_appSettings.JWTSettings.SecurityKey))
+            }, out SecurityToken validatedToken);
+        }
+        catch (Exception)
+        {
+            return false;
+        }
+        return true;
+    }
+
     public async void NotifyUserLogout()
     {
+        if (!_firsloadState.IsFirstLoaded)
+        {
+            return;
+        }
         await RemoveCurrentUser();
         var authState = Task.FromResult(_anonymous);
         NotifyAuthenticationStateChanged(authState);
